requirements.txt
gradio
networkx
pydantic
python-dotenv
matplotlib
httpx
typing-extensions
rich
pandas
numpy
google-generativeai
openai
langchain
langchain-openai
langchain-ollama
matplotlib-venn
python-dotenv
